requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv
Pillow
psycopg2
six
